Feature/option feature

+++ src/controller/event.js
@@ -1,13 +1,22 @@
 const repo = require("../repository")
 const utils = require("../utils")
 const { CustomError } = require("../config/error")
-const { coverImagePath, eventImagePath } = require("../constant/cludinaryPath")
 const fs = require("fs")
 const { FACILITY_LIST } = require("../constant")
+const { ROLE } = require("../constant/enum")
 
 exports.getAll = utils.catchError(async (req, res, next) => {
     const allEvent = await repo.event.getAll()
     res.status(200).json(allEvent)
+})
+
+exports.getAllUpcomimng = utils.catchError(async (req, res, next) => {
+    const today = new Date();
+    const targetDate = new Date(today)
+    targetDate.setDate(today.getDate() + 30)
+
+    const AllUpcomimng = await repo.event.getAllUpcomimng(targetDate)
+    res.status(200).json(AllUpcomimng)
 })
 
 exports.getEvent = utils.catchError(async (req, res, next) => {
@@ -18,10 +27,10 @@ exports.getEvent = utils.catchError(async (req, res, next) => {
 })
 
 exports.getinRange = utils.catchError(async (req, res, next) => {
-    const {firstDay, lastDay} = req.body
+    const { firstDay, lastDay } = req.body
     const a = req.body
 
-    const events = await repo.event.getRangeEvent(firstDay,lastDay)
+    const events = await repo.event.getRangeEvent(firstDay, lastDay)
     res.status(200).json(events)
 })
 
@@ -39,6 +48,7 @@ exports.createEvent = utils.catchError(async (req, res, next) => {
         districtId,
         subDistrictId,
         address,
+        address2,
         lat,
         long,
         ...eventData
@@ -55,11 +65,14 @@ exports.createEvent = utils.catchError(async (req, res, next) => {
     }
 
     // UPLOAD coverImage to Cloudinary
-    const coverImageUrl = await utils.cloudinary.uploadImage(coverImage[0].path, coverImagePath)
+    const coverImageUrl = await utils.cloudinary.uploadImage(coverImage[0].path)
     eventData.coverImage = coverImageUrl.secure_url
 
     // CREATE event
     const event = await repo.event.createEvent(eventData)
+
+    // Delete image in public folder
+    fs.unlink(coverImage[0].path, () => {})
 
     // CREATE Facility
     const facilityData = { petFriendly, wifi, entranceFee, prayerRoom, toilet, parking, medicalService, food }
@@ -74,31 +87,30 @@ exports.createEvent = utils.catchError(async (req, res, next) => {
     await repo.event.createFacility(facilityData)
 
     // CREATE event address
-    const eventAdressData = { provinceId, districtId, subDistrictId, address, lat, long, eventId: event.id }
+    const eventAdressData = { provinceId, districtId, address2, subDistrictId, address, lat, long, eventId: event.id }
     for (const key in eventAdressData) {
-        if (key !== "address") {
+        if (key !== "address" && key !== "address2") {
             eventAdressData[key] = +eventAdressData[key]
         }
     }
     await repo.event.createEventAddess(eventAdressData)
 
     // UPLOAD eventImage to Cloudinary
-    const eventImageData = []
-    for (file of image) {
-        const { path } = file
-        const eventImageUrl = await utils.cloudinary.uploadImage(path, eventImagePath)
-        eventImageData.push({ eventId: event.id, image: eventImageUrl.secure_url })
+    if (image) {
+        const eventImageData = []
+        for (file of image) {
+            const { path } = file
+            const eventImageUrl = await utils.cloudinary.uploadImage(path)
+            eventImageData.push({ eventId: event.id, image: eventImageUrl.secure_url })
+        }
+        // CREATE eventImage
+        await repo.eventImage.createEventImages(eventImageData)
+        for (file of image) {
+            const { path } = file
+            fs.unlink(path, () => {})
+        }
     }
 
-    // CREATE eventImage
-    await repo.eventImage.createEventImages(eventImageData)
-
-    // Delete image in public folder
-    fs.unlink(coverImage[0].path, () => {})
-    for (file of image) {
-        const { path } = file
-        fs.unlink(path, () => {})
-    }
     res.status(200).json(event.id)
 })
 
@@ -114,7 +126,7 @@ module.exports.getAllInScope = utils.catchError(async (req, res, next) => {
 })
 
 module.exports.getFilteredEvent = utils.catchError(async (req, res, next) => {
-    // console.log(req.body);
+    const today = new Date();
     const data = req.body
 
     const where = {}
@@ -142,7 +154,9 @@ module.exports.getFilteredEvent = utils.catchError(async (req, res, next) => {
             where.EventFacility[value] = true
         }
     }
-
+    where.startDate = {
+        gt: today,
+    }
     // console.log(where);
 
     const events = await repo.event.getFilteredEvent(where)
@@ -165,6 +179,7 @@ module.exports.updateEvent = utils.catchError(async (req, res, next) => {
         districtId,
         subDistrictId,
         address,
+        address2,
         lat,
         long,
         ...eventData
@@ -195,11 +210,16 @@ module.exports.updateEvent = utils.catchError(async (req, res, next) => {
     }
 
     // UPLOAD coverImage to Cloudinary
-    const coverImageUrl = await utils.cloudinary.uploadImage(coverImage.path, coverImagePath)
-    eventData.coverImage = coverImageUrl.secure_url
+    if (coverImage) {
+        const coverImageUrl = await utils.cloudinary.uploadImage(coverImage.path)
+        eventData.coverImage = coverImageUrl.secure_url
+        fs.unlink(coverImage.path, () => {})
+    }
 
     // UPDATE event
-    eventData.categoryId = +eventData.categoryId
+    if (eventData.categoryId) {
+        eventData.categoryId = +eventData.categoryId
+    }
     const event = await repo.event.updateEvent({ id: +eventId }, eventData)
 
     // UPDATE facility
@@ -212,12 +232,11 @@ module.exports.updateEvent = utils.catchError(async (req, res, next) => {
         }
     }
     await repo.event.updateFacility({ eventId: +eventId }, facilityData)
-    fs.unlink(coverImage.path, () => {})
 
     // UPDATE Event address
-    const eventAdressData = { provinceId, districtId, subDistrictId, address, lat, long, eventId: event.id }
+    const eventAdressData = { provinceId, districtId, subDistrictId, address2, address, lat, long, eventId: event.id }
     for (const key in eventAdressData) {
-        if (key !== "address" && eventAdressData[key]) {
+        if ((key !== "address" && key !== "address2") && eventAdressData[key]) {
             eventAdressData[key] = +eventAdressData[key]
         }
     }
@@ -246,11 +265,57 @@ module.exports.deleteEvent = utils.catchError(async (req, res, next) => {
     await repo.eventImage.deleteEventImages({ eventId: +eventId })
     await repo.event.deleteEventAddess({ eventId: +eventId })
     await repo.event.deleteHighlightEvent({ eventId: +eventId })
-    await repo.event.deleteReport({ eventId: +eventId })
     await repo.reminder.deleteReminderByEventId({ eventId: +eventId })
     await repo.event.deleteEventFeedback({ eventId: +eventId })
     await repo.event.deleteFacility({ eventId: +eventId })
     await repo.event.deleteEvent({ id: +eventId })
 
     res.status(200).json({ message: "Delete success" })
+})
+
+// =========================================== HighLight ====================================== //
+
+module.exports.createHighlight = utils.catchError(async (req,res, next) => {
+    const { id } = req.user
+    const   {eventId}  = req.body
+
+    const admin = await repo.user.getUser({id})
+    if(admin.role !== ROLE.ADMIN){
+        throw new CustomError("No authorize to do it", "Invalid Authorization", 401)
+    }
+
+    await repo.event.createHighlight({eventId})
+
+
+    res.status(200).json({message: "add Success"})
+})
+
+module.exports.deleteHighlight = utils.catchError(async (req,res, next) => {
+    const { id } = req.user
+    const   {eventId}  = req.body
+
+    const admin = await repo.user.getUser({id})
+    if(admin.role !== ROLE.ADMIN){
+        throw new CustomError("No authorize to do it", "Invalid Authorization", 401)
+    }
+
+    await repo.event.deleteHighlight({eventId})
+
+
+    res.status(200).json({message: "Remove Success"})
+})
+
+// =========================================== feedBack ====================================== //
+
+module.exports.createFeedback = utils.catchError(async (req,res, next) => {
+    const { id } = req.user
+    const {eventId}  = req.params
+    const {content, isLike} = req.body
+    const feedBackData = {userId:+id ,eventId:+eventId ,content,isLike}
+
+
+    await repo.event.createFeedback(feedBackData)
+
+
+    res.status(200).json({message: "create Feedback Success"})
 })
